+++ assets/javascripts/kitten/components/layout/dashboard-layout/stories.js
@@ -484,8 +484,13 @@ const SiteHeaderComponent = () => {
 
       <HeaderNav.Right>
         <HeaderNav.Logged>
-          <HeaderNav.UserMenu dropdownContentWidth={userMenuWidth} hasArrow>
-            <HeaderNav.UserMenu.Button>
+          <HeaderNav.UserMenu dropdownContentWidth={userMenuWidth}>
+            <HeaderNav.UserMenu.Button
+              hasArrow
+              backgroundColor={COLORS.background1}
+              backgroundColorHover={COLORS.line1}
+              backgroundColorActive={COLORS.line1}
+            >
               <AvatarWithTextAndBadge>
                 <AvatarWithTextAndBadge.Image src="https://via.placeholder.com/40x40.png">
                   <AvatarWithTextAndBadge.Badge>2</AvatarWithTextAndBadge.Badge>