+++ CHANGELOG.md
@@ -5,7 +5,14 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: `HeaderNav`: Add `size` prop, accepts `regular` and `small`.
-- Feature: `HeaderNav.UserMenu`: Add `hasArrow` boolean prop.
+- Feature: `HeaderNav.UserMenu.Button`: Add `hasArrow` boolean prop.
+- Feature: `HeaderNav.UserMenu.Button`: Add the following color props:
+  - `backgroundColor`
+  - `backgroundColorHover`
+  - `backgroundColorActive`
+  - `color`
+  - `colorHover`
+  - `colorActive`
 - Feature: `DashboardLayout`: Add `DashboardLayout.SiteHeader` component.
 
 ## [3.9.0] - 2021-05-06